Compute air potential temperature from in-situ temperature

Potential temperature is computed as in Chapter 3 of Reference 1, i.e. as the reverse of the formula used for T2theta().

Usage

T2theta(temperature, pressure, Rd = 287.04, Cp = 1005.7)

Arguments

temperature

numeric value indicating in-situ temperature in degrees C.

pressure

numeric value indicating pressure in mbar (equivalently, in hPa).

Rd, Cp

dry-air gas constant and mass-specific heat capacity, with the default values as in Reference 2.

Value

T2theta returns potential temperature, in degrees Celcius.

References

Wallace, John M., and Peter Victor Hobbs. Atmospheric Science: An Introductory Survey. 2nd ed. International Geophysics Series, volume 92. Amsterdam ; Boston: Elsevier Academic Press, 2006.

https://github.com/cwebster2/pyMeteo

Author

Dan Kelley, with the default values for Rd and Cp taken from Reference 1.