Download atmospheric sounding data

The file is downloaded from Reference 1, by constructing a URL that queries the server for a particular station and time. The construction of that URL is based on reverse-engineering the URLs that get created by clicking on a map. This means that downloadAtmosphericSounding is brittle, and will fail if the server setup changes. Unfortunately, this guessing game seems to be the best we can do, since the website does not provide information on the URL construction, let alone an API.

Usage

downloadAtmosphericSounding(
  date = Sys.Date() - 1,
  region = "naconf",
  stationName = "Shearwater"
)

Arguments

date

A string for a date, in yyyy-mm-dd format, or some other form that as.POSIXlt() can convert to a date or datetime. If not provided, the previous day is used (because an early-morning call to this function might yield a failed download).

region

a character value indicating the region. This defaults to "naconf.

stationName

a character value indicating the station name, which the present function converts to a station number by stationNumber().

References

https://weather.uwyo.edu/upperair/sounding.html

Author

Dan Kelley